requirements.txt
Flask
google-generativeai
python-dotenv
speechrecognition
pyttsx3
faiss-cpu
sentence-transformers
